requirements.txt
moviepy==1.0.3
google-cloud-speech==2.10.0
google-cloud-translate==3.7.3
google-cloud-vision==2.7.1
python-docx==0.8.11